后端代码/requirements.txt
aliyun_python_sdk_core==2.13.36
caching==0.1.dev8
Flask==2.2.5
Flask_Cors==3.0.10
flask_sqlalchemy==3.0.5
PyJWT==2.7.0
PyJWT==2.7.0
redis==4.5.5
Requests==2.31.0
SQLAlchemy==2.0.16
